+++ backend/src/main/java/com/nossagrana/backend/dto/CategoriaResponse.java
@@ -4,17 +4,18 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class CategoriaResponse {
-    
+
     private Long id;
     private String nome;
     private String icone;
     private String cor;
     private Boolean ativa;
-    private Double orcamentoMensal;
+    private BigDecimal orcamentoMensal;
 }
